Move KotlinCompilerPluginArgsEncoder to toolchain. Temporary placement, new location TBD
